+++ src/containers/Auth/hooks/useMe.ts
@@ -1,8 +1,7 @@
 // deps
-import { useContext, useCallback } from 'react'
+import React from 'react'
 import { useLocation } from 'react-router-dom'
 // model
-import { EAuthContextActions, IAuthReducerAction } from '_/model/context/auth'
 import { IQueryMeResponse } from '_/model/queries/auth'
 // helpers
 import { SnackbarContext } from '_/context'
@@ -10,31 +9,20 @@ import { queryMe } from '_/gql/queries'
 import { useFetch } from '_/utils/hooks'
 import { useLogout } from '_/containers/Auth/hooks/useLogout'
 
-export interface IUseCheckAuthorizationProps {
-  dispatch: (value: IAuthReducerAction) => void
-}
-
-export const useCheckAuthorization = ({
-  dispatch,
-}: IUseCheckAuthorizationProps) => {
+export const useCheckAuthorization = () => {
   const location = useLocation()
-  const { setSnackbarState } = useContext(SnackbarContext)
+  const { setSnackbarState } = React.useContext(SnackbarContext)
   const handleFetch = useFetch()
   const handleLogout = useLogout()
 
-  return useCallback(async () => {
+  const handleCheckAuthorization = async () => {
     try {
       const res = await handleFetch(queryMe())
-      const { errors, data }: IQueryMeResponse = await res.json()
-      if (data?.me) {
-        dispatch({
-          type: EAuthContextActions.LOGIN,
-          payload: {
-            userCredentials: data.me,
-          },
-        })
-      } else if (
-        errors?.some((error) => error.statusCode === 401) &&
+      const { errors }: IQueryMeResponse = await res.json()
+      if (
+        errors?.some(
+          (error: Record<'statusCode', number>) => error.statusCode === 401,
+        ) &&
         location.pathname !== '/auth'
       ) {
         handleLogout()
@@ -51,5 +39,7 @@ export const useCheckAuthorization = ({
         open: true,
       })
     }
-  }, [dispatch, handleFetch, handleLogout, location.pathname, setSnackbarState])
+  }
+
+  return [handleCheckAuthorization]
 }