+++ src/containers/Employee/hooks/useGetDepartments.ts
@@ -9,7 +9,7 @@ import {
 } from '_/model/context/catalogs'
 // helpers
 import { useFetch } from '_/utils/hooks'
-import { getDepartments } from '_/gql/queries'
+import { queryDepartments } from '_/gql/queries'
 import { fetchResponseCheck } from '_/utils/auth'
 
 export interface IUseGetDepartments {
@@ -17,11 +17,11 @@ export interface IUseGetDepartments {
 }
 
 export const useGetDepartments = ({ dispatch }: IUseGetDepartments) => {
-  const [handleFetch] = useFetch()
+  const handleFetch = useFetch()
   const handleGetDepartments = React.useCallback(async () => {
     dispatch({ type: EActionTypes.LOADING, payload: { loading: true } })
     try {
-      const res = await handleFetch(getDepartments())
+      const res = await handleFetch(queryDepartments())
       fetchResponseCheck(res?.status)
       const {
         data: { departments },
@@ -35,6 +35,6 @@ export const useGetDepartments = ({ dispatch }: IUseGetDepartments) => {
       dispatch({ type: EActionTypes.ERROR, payload: { error } })
     }
     dispatch({ type: EActionTypes.LOADING, payload: { loading: false } })
-  }, [handleFetch])
+  }, [dispatch, handleFetch])
   return [handleGetDepartments]
 }