+++ src/gql/queries/index.ts
@@ -16,17 +16,16 @@ import {
   titleFragment,
 } from '_/gql/fragments'
 
-export const loginQuery = ({ email, password }: UserInput): IQueryProps => ({
+export const queryLogin = ({ email, password }: UserInput): IQueryProps => ({
   query: `
-        query loginQuery($email: String!, $password: String!) {
+        query login($email: String!, $password: String!) {
           login(
             email: $email, 
             password: $password){
             userCredentials{
-              id
+              _id
               email
             }
-            token
           } 
         }
       `,
@@ -36,7 +35,15 @@ export const loginQuery = ({ email, password }: UserInput): IQueryProps => ({
   },
 })
 
-export const getEmployees = ({
+export const queryLogout = () => ({
+  query: `
+        query logout {
+          logout
+        }
+  `,
+})
+
+export const queryEmployees = ({
   limit,
   offset,
 }: GetEmployeesInput): IQueryProps => ({
@@ -60,7 +67,7 @@ export const getEmployees = ({
   },
 })
 
-export const getEmployee = ({ id }: GetEmployeeInput): IQueryProps => ({
+export const queryEmployee = ({ id }: GetEmployeeInput): IQueryProps => ({
   query: `
     query employee($id: ID!){
       employee(input:{id: $id}){
@@ -71,7 +78,7 @@ export const getEmployee = ({ id }: GetEmployeeInput): IQueryProps => ({
   variables: { id },
 })
 
-export const getMe = (): IQueryProps => ({
+export const queryMe = (): IQueryProps => ({
   query: `
     query me {
       ${userCredentials}
@@ -79,7 +86,7 @@ export const getMe = (): IQueryProps => ({
   `,
 })
 
-export const getDepartments = (): IQueryProps => ({
+export const queryDepartments = (): IQueryProps => ({
   query: `
     query departments{
       departments{
@@ -89,7 +96,7 @@ export const getDepartments = (): IQueryProps => ({
   `,
 })
 
-export const getTitles = (): IQueryProps => ({
+export const queryTitles = (): IQueryProps => ({
   query: `
     query titles {
       titles{
@@ -99,7 +106,7 @@ export const getTitles = (): IQueryProps => ({
   `,
 })
 
-export const getGenders = (): IQueryProps => ({
+export const queryGenders = (): IQueryProps => ({
   query: `
     query genders {
       genders{