+++ CHANGELOG.md
@@ -1,7 +1,16 @@
-# v0.1.0 (Sun Mar 21 2021)
+# v0.1.2
+## Authorization improvements
+- Add cookies authorization
+- Add husky pre-commit check
+- Add readme
 
+
+# v0.1.1
+## Authorization improvements
+- Add proxy for BE queries
+
+# v0.1.0
 ## Init release
-
 - Add pages Login, Main, Employees, Employee, About
 - Add react based data management
 - Add graphql schema import from BE